Support loading environment from external location
Adds support for an $ENV_URL to load environment variables from an
external source (e.g. S3) with which the rails application is then
loaded.

This would enable us to store application configuration in external
services such as S3 (combined with appropriate access control policies)
and load this configuration when the application server is started.

Would still need to add additional error-handling in case of failed
requests or syntactically invalid content.